connectors/webhook/src/requirements.txt
pycti
requests
pyyaml
